Convert Huawei GDPR export to GPX
Some R code snippet to convert the GDPR export from Huawei to individual GPX files. You will need to make some changes to the code for it to run on your computer. Not all the source and library commands at the top are needed, so please remove whatever is not required.